Optimized data structures and refactored code
Significant changes include:
- Replaced std::unordered_map with boost::container::flat_map for better performance
- Refactored AllowedRelationships class to improve efficiency and readability
- Changed uid type from int to unsigned short in Infection class
- Updated parseAllowedParentsFromJSON function to use shared_ptr for AllowedRelationships object
- Modified RandomAllelesBitSetSampler4 class to use shared_ptr for AllowedRelationships object
- Adjusted ModelNine implementation to accommodate the aforementioned changes

+++ src/core/containers/Infection.h
@@ -38,7 +38,7 @@ namespace transmission_nets::core::containers {
         explicit Infection(std::string id, float observationTime, bool symptomatic = true);
 
         Infection(const Infection& other, const std::string& id = "", bool retain_alleles = true) {
-            static int uid = 0;
+            static unsigned short uid = 0;
             uid_ = uid++;
             // Copy constructor -- create a new infection from an existing one using the latent genetics.
             if (id.empty()) {
@@ -165,7 +165,7 @@ namespace transmission_nets::core::containers {
             return id_;
         }
 
-        [[nodiscard]] int uid() const {
+        [[nodiscard]] unsigned short uid() const {
             return uid_;
         }
